+++ src/main/java/ee/carlrobert/codegpt/CodeGPTKeys.java
@@ -2,7 +2,10 @@ package ee.carlrobert.codegpt;
 
 import com.intellij.openapi.util.Key;
 import ee.carlrobert.codegpt.predictions.CodeSuggestionDiffViewer;
+import ee.carlrobert.codegpt.toolwindow.chat.editor.ToolWindowEditorFileDetails;
 import ee.carlrobert.llm.client.codegpt.CodeGPTUserDetails;
+import ee.carlrobert.service.NextEditResponse;
+import ee.carlrobert.service.PartialCodeCompletionResponse;
 
 public class CodeGPTKeys {
 
@@ -18,4 +21,10 @@ public class CodeGPTKeys {
       Key.create("codegpt.isPromptTextFieldDocument");
   public static final Key<CodeSuggestionDiffViewer> EDITOR_PREDICTION_DIFF_VIEWER =
       Key.create("codegpt.editorPredictionDiffViewer");
+  public static final Key<PartialCodeCompletionResponse> REMAINING_CODE_COMPLETION =
+      Key.create("codegpt.remainingCodeCompletion");
+  public static final Key<NextEditResponse> REMAINING_PREDICTION_RESPONSE =
+      Key.create("codegpt.remainingPredictionResponse");
+  public static final Key<ToolWindowEditorFileDetails> TOOLWINDOW_EDITOR_FILE_DETAILS =
+      Key.create("proxyai.toolwindowEditorFileDetails");
 }