+++ src/main/java/ee/carlrobert/codegpt/toolwindow/chat/ui/ChatMessageResponseBody.java
@@ -1,6 +1,5 @@
 package ee.carlrobert.codegpt.toolwindow.chat.ui;
 
-import static ee.carlrobert.codegpt.toolwindow.chat.StreamResponseType.CODE;
 import static ee.carlrobert.codegpt.util.MarkdownUtil.convertMdToHtml;
 import static java.lang.String.format;
 import static javax.swing.event.HyperlinkEvent.EventType.ACTIVATED;
@@ -21,11 +20,11 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.io.FileUtil;
 import com.intellij.openapi.vfs.LocalFileSystem;
 import com.intellij.openapi.vfs.VirtualFile;
+import com.intellij.ui.AnimatedIcon;
 import com.intellij.ui.PopupHandler;
 import com.intellij.ui.components.JBLabel;
 import com.intellij.util.ui.JBUI;
-import com.vladsch.flexmark.ast.FencedCodeBlock;
-import com.vladsch.flexmark.parser.Parser;
+import com.intellij.util.ui.components.BorderLayoutPanel;
 import ee.carlrobert.codegpt.CodeGPTBundle;
 import ee.carlrobert.codegpt.Icons;
 import ee.carlrobert.codegpt.actions.ActionType;
@@ -38,24 +37,27 @@ import ee.carlrobert.codegpt.settings.GeneralSettings;
 import ee.carlrobert.codegpt.settings.GeneralSettingsConfigurable;
 import ee.carlrobert.codegpt.settings.service.ServiceType;
 import ee.carlrobert.codegpt.telemetry.TelemetryAction;
-import ee.carlrobert.codegpt.toolwindow.chat.StreamParser;
-import ee.carlrobert.codegpt.toolwindow.chat.ThinkingOutputParser;
 import ee.carlrobert.codegpt.toolwindow.chat.editor.ResponseEditorPanel;
 import ee.carlrobert.codegpt.toolwindow.chat.editor.actions.CopyAction;
+import ee.carlrobert.codegpt.toolwindow.chat.parser.CompleteOutputParser;
+import ee.carlrobert.codegpt.toolwindow.chat.parser.StreamOutputParser;
+import ee.carlrobert.codegpt.toolwindow.chat.parser.StreamParseResponse;
+import ee.carlrobert.codegpt.toolwindow.chat.parser.StreamParseResponse.StreamResponseType;
 import ee.carlrobert.codegpt.toolwindow.ui.ResponseBodyProgressPanel;
 import ee.carlrobert.codegpt.toolwindow.ui.WebpageList;
 import ee.carlrobert.codegpt.ui.ThoughtProcessPanel;
 import ee.carlrobert.codegpt.ui.UIUtil;
 import ee.carlrobert.codegpt.util.EditorUtil;
-import ee.carlrobert.codegpt.util.MarkdownUtil;
 import java.awt.BorderLayout;
 import java.util.Objects;
 import java.util.stream.Stream;
 import javax.swing.BoxLayout;
 import javax.swing.DefaultListModel;
 import javax.swing.JEditorPane;
+import javax.swing.JLabel;
 import javax.swing.JPanel;
 import javax.swing.JTextPane;
+import javax.swing.event.HyperlinkListener;
 import org.jetbrains.annotations.NotNull;
 
 public class ChatMessageResponseBody extends JPanel {
@@ -64,18 +66,28 @@ public class ChatMessageResponseBody extends JPanel {
 
   private final Project project;
   private final Disposable parentDisposable;
-  private final StreamParser streamParser;
-  private final ThinkingOutputParser thinkingOutputParser;
+  private final StreamOutputParser streamOutputParser;
   private final boolean readOnly;
   private final DefaultListModel<WebSearchEventDetails> webpageListModel = new DefaultListModel<>();
   private final WebpageList webpageList = new WebpageList(webpageListModel);
   private final ResponseBodyProgressPanel progressPanel = new ResponseBodyProgressPanel();
+  private final JPanel loadingLabel = createLoadingPanel();
+  private final JPanel contentPanel = new JPanel();
   private ResponseEditorPanel currentlyProcessedEditorPanel;
   private JEditorPane currentlyProcessedTextPane;
   private JPanel webpageListPanel;
 
+  private JPanel createLoadingPanel() {
+    return new BorderLayoutPanel()
+        .addToLeft(new JBLabel(
+            CodeGPTBundle.get("toolwindow.chat.loading"),
+            new AnimatedIcon.Default(),
+            JLabel.LEFT))
+        .withBorder(JBUI.Borders.empty(4, 0));
+  }
+
   public ChatMessageResponseBody(Project project, Disposable parentDisposable) {
-    this(project, false, false, false, parentDisposable);
+    this(project, false, false, false, false, parentDisposable);
   }
 
   public ChatMessageResponseBody(
@@ -83,31 +95,39 @@ public class ChatMessageResponseBody extends JPanel {
       boolean readOnly,
       boolean webSearchIncluded,
       boolean withProgress,
+      boolean withLoading,
       Disposable parentDisposable) {
     this.project = project;
     this.parentDisposable = parentDisposable;
-    this.streamParser = new StreamParser();
-    this.thinkingOutputParser = new ThinkingOutputParser();
+    this.streamOutputParser = new StreamOutputParser();
     this.readOnly = readOnly;
-    setLayout(new BoxLayout(this, BoxLayout.Y_AXIS));
+
+    setLayout(new BorderLayout());
     setOpaque(false);
+
+    contentPanel.setLayout(new BoxLayout(contentPanel, BoxLayout.Y_AXIS));
+    contentPanel.setOpaque(false);
+    add(contentPanel, BorderLayout.NORTH);
+
+    loadingLabel.setVisible(withLoading);
+    add(loadingLabel, BorderLayout.SOUTH);
 
     if (GeneralSettings.getSelectedService() == ServiceType.CODEGPT) {
       if (withProgress) {
-        add(progressPanel);
+        contentPanel.add(progressPanel);
       }
 
       if (webSearchIncluded) {
         webpageListPanel = createWebpageListPanel(webpageList);
-        add(webpageListPanel);
+        contentPanel.add(webpageListPanel);
       }
     }
   }
 
   public ChatMessageResponseBody withResponse(@NotNull String response) {
     try {
-      for (var message : MarkdownUtil.splitCodeBlocks(response)) {
-        processResponse(message, message.startsWith("```"), false);
+      for (var item : new CompleteOutputParser().parse(response)) {
+        processResponse(item, false);
 
         currentlyProcessedTextPane = null;
         currentlyProcessedEditorPanel = null;
@@ -118,81 +138,47 @@ public class ChatMessageResponseBody extends JPanel {
     return this;
   }
 
+  public void stopLoading() {
+    loadingLabel.setVisible(false);
+  }
+
   public void updateMessage(String partialMessage) {
     if (partialMessage.isEmpty()) {
       return;
     }
 
-    var processedPartialMessage = processThinkingOutput(partialMessage);
-    if (processedPartialMessage.isEmpty()) {
-      return;
-    }
-
-    for (var item : streamParser.parse(processedPartialMessage)) {
-      processResponse(item.response(), CODE.equals(item.type()), true);
+    var parsedResponse = streamOutputParser.parse(partialMessage);
+    for (StreamParseResponse item : parsedResponse) {
+      processResponse(item, true);
     }
   }
 
   public void displayMissingCredential() {
-    ApplicationManager.getApplication().invokeLater(() -> {
-      var message = "API key not provided. Open <a href=\"#\">Settings</a> to set one.";
-      currentlyProcessedTextPane.setText(
-          format("<html><p style=\"margin-top: 4px; margin-bottom: 8px;\">%s</p></html>", message));
-      currentlyProcessedTextPane.addHyperlinkListener(e -> {
-        if (e.getEventType() == ACTIVATED) {
-          ShowSettingsUtil.getInstance()
-              .showSettingsDialog(project, GeneralSettingsConfigurable.class);
-        }
-      });
-      hideCaret();
-
-      if (webpageListPanel != null) {
-        webpageListPanel.setVisible(false);
+    String message = "API key not provided. Open <a href=\"#\">Settings</a> to set one.";
+    displayErrorMessage(message, e -> {
+      if (e.getEventType() == ACTIVATED) {
+        ShowSettingsUtil.getInstance()
+            .showSettingsDialog(project, GeneralSettingsConfigurable.class);
       }
     });
   }
 
   public void displayQuotaExceeded() {
-    ApplicationManager.getApplication().invokeLater(() -> {
-      currentlyProcessedTextPane.setText("<html>"
-          + "<p style=\"margin-top: 4px; margin-bottom: 8px;\">"
-          + "You exceeded your current quota, please check your plan and billing details, "
-          + "or <a href=\"#CHANGE_PROVIDER\">change</a> to a different LLM provider.</p>"
-          + "</html>");
-
-      currentlyProcessedTextPane.addHyperlinkListener(e -> {
-        if (e.getEventType() == ACTIVATED) {
-          ShowSettingsUtil.getInstance()
-              .showSettingsDialog(project, GeneralSettingsConfigurable.class);
-          TelemetryAction.IDE_ACTION.createActionMessage()
-              .property("action", ActionType.CHANGE_PROVIDER.name())
-              .send();
-        }
-      });
-      hideCaret();
-
-      if (webpageListPanel != null) {
-        webpageListPanel.setVisible(false);
+    String message = "You exceeded your current quota, please check your plan and billing details, "
+        + "or <a href=\"#CHANGE_PROVIDER\">change</a> to a different LLM provider.";
+    displayErrorMessage(message, e -> {
+      if (e.getEventType() == ACTIVATED) {
+        ShowSettingsUtil.getInstance()
+            .showSettingsDialog(project, GeneralSettingsConfigurable.class);
+        TelemetryAction.IDE_ACTION.createActionMessage()
+            .property("action", ActionType.CHANGE_PROVIDER.name())
+            .send();
       }
     });
   }
 
   public void displayError(String message) {
-    ApplicationManager.getApplication().invokeLater(() -> {
-      var errorText = format(
-          "<html><p style=\"margin-top: 4px; margin-bottom: 8px;\">%s</p></html>",
-          message);
-      if (currentlyProcessedTextPane == null) {
-        add(createTextPane(errorText));
-      } else {
-        currentlyProcessedTextPane.setText(errorText);
-      }
-      hideCaret();
-
-      if (webpageListPanel != null) {
-        webpageListPanel.setVisible(false);
-      }
-    });
+    displayErrorMessage(message, null);
   }
 
   public void handleCodeGPTEvent(CodeGPTEvent codegptEvent) {
@@ -229,9 +215,10 @@ public class ChatMessageResponseBody extends JPanel {
   }
 
   public void clear() {
-    removeAll();
+    contentPanel.removeAll();
+    streamOutputParser.clear();
+    loadingLabel.setVisible(false);
 
-    streamParser.clear();
     // TODO: First message might be code block
     prepareProcessingText(true);
     currentlyProcessedTextPane.setText(
@@ -241,55 +228,85 @@ public class ChatMessageResponseBody extends JPanel {
     revalidate();
   }
 
-  private String processThinkingOutput(String partialMessage) {
-    var processedChunk = thinkingOutputParser.processChunk(partialMessage);
-    var thoughtProcessPanel = getExistingThoughtProcessPanel();
-
-    if (thinkingOutputParser.isThinking()) {
-      progressPanel.setVisible(false);
-
-      if (thoughtProcessPanel == null) {
-        thoughtProcessPanel = new ThoughtProcessPanel();
-        add(thoughtProcessPanel);
-      } else {
-        thoughtProcessPanel.updateText(thinkingOutputParser.getThoughtProcess());
+  private void displayErrorMessage(String message, HyperlinkListener hyperlinkListener) {
+    ApplicationManager.getApplication().invokeLater(() -> {
+      if (loadingLabel.isVisible()) {
+        loadingLabel.setVisible(false);
       }
-    }
+      if (webpageListPanel != null) {
+        webpageListPanel.setVisible(false);
+      }
 
-    if (thoughtProcessPanel != null && thinkingOutputParser.isFinished()) {
-      thoughtProcessPanel.setFinished();
-    }
+      if (currentlyProcessedTextPane == null) {
+        currentlyProcessedTextPane = createTextPane("");
+        contentPanel.add(currentlyProcessedTextPane);
+      }
 
-    return processedChunk;
+      String formattedMessage = format(
+          "<html><p style=\"margin-top: 4px; margin-bottom: 8px;\">%s</p></html>", message);
+      currentlyProcessedTextPane.setVisible(true);
+      currentlyProcessedTextPane.setText(formattedMessage);
+
+      if (hyperlinkListener != null) {
+        for (HyperlinkListener listener : currentlyProcessedTextPane.getHyperlinkListeners()) {
+          currentlyProcessedTextPane.removeHyperlinkListener(listener);
+        }
+        currentlyProcessedTextPane.addHyperlinkListener(hyperlinkListener);
+      }
+
+      hideCaret();
+
+      revalidate();
+      repaint();
+    });
+  }
+
+  private void processThinkingOutput(String thoughtProcess) {
+    progressPanel.setVisible(false);
+
+    var thoughtProcessPanel = getExistingThoughtProcessPanel();
+    if (thoughtProcessPanel == null) {
+      thoughtProcessPanel = new ThoughtProcessPanel();
+      thoughtProcessPanel.updateText(thoughtProcess);
+      contentPanel.add(thoughtProcessPanel);
+    } else {
+      thoughtProcessPanel.updateText(thoughtProcess);
+    }
   }
 
   private ThoughtProcessPanel getExistingThoughtProcessPanel() {
-    return (ThoughtProcessPanel) Stream.of(getComponents())
+    return (ThoughtProcessPanel) Stream.of(contentPanel.getComponents())
         .filter(it -> it instanceof ThoughtProcessPanel)
         .findFirst()
         .orElse(null);
   }
 
-  private void processResponse(String markdownInput, boolean codeResponse, boolean caretVisible) {
-    if (codeResponse) {
-      processCode(markdownInput);
+  private void processResponse(StreamParseResponse item, boolean caretVisible) {
+    if (item.getType() == StreamResponseType.THINKING) {
+      processThinkingOutput(item.getContent());
+      return;
+    }
+
+    var thoughtProcessPanel = getExistingThoughtProcessPanel();
+    if (thoughtProcessPanel != null && !thoughtProcessPanel.isFinished()) {
+      thoughtProcessPanel.setFinished();
+    }
+
+    if (item.getType() == StreamResponseType.CODE_CONTENT
+        || item.getType() == StreamResponseType.CODE_HEADER) {
+      processCode(item);
     } else {
-      processText(markdownInput, caretVisible);
+      processText(item.getContent(), caretVisible);
     }
   }
 
-  private void processCode(String markdownCode) {
-    var document = Parser.builder().build().parse(markdownCode);
-    var child = document.getChildOfType(FencedCodeBlock.class);
-    if (child != null) {
-      var codeBlock = ((FencedCodeBlock) child);
-      var code = codeBlock.getContentChars().toString();
-      if (!code.isEmpty()) {
-        if (currentlyProcessedEditorPanel == null) {
-          prepareProcessingCode(code, codeBlock.getInfo().toString());
-        }
-        EditorUtil.updateEditorDocument(currentlyProcessedEditorPanel.getEditor(), code);
+  private void processCode(StreamParseResponse item) {
+    var content = item.getContent();
+    if (!content.isEmpty()) {
+      if (currentlyProcessedEditorPanel == null) {
+        prepareProcessingCode(item);
       }
+      EditorUtil.updateEditorDocument(currentlyProcessedEditorPanel.getEditor(), content);
     }
   }
 
@@ -304,15 +321,15 @@ public class ChatMessageResponseBody extends JPanel {
   private void prepareProcessingText(boolean caretVisible) {
     currentlyProcessedEditorPanel = null;
     currentlyProcessedTextPane = createTextPane("", caretVisible);
-    add(currentlyProcessedTextPane);
+    contentPanel.add(currentlyProcessedTextPane);
   }
 
-  private void prepareProcessingCode(String code, String markdownLanguage) {
+  private void prepareProcessingCode(StreamParseResponse item) {
     hideCaret();
     currentlyProcessedTextPane = null;
     currentlyProcessedEditorPanel =
-        new ResponseEditorPanel(project, code, markdownLanguage, readOnly, parentDisposable);
-    add(currentlyProcessedEditorPanel);
+        new ResponseEditorPanel(project, item, readOnly, parentDisposable);
+    contentPanel.add(currentlyProcessedEditorPanel);
   }
 
   private void displayWebSearchItem(WebSearchEventDetails details) {