+++ src/main/kotlin/ee/carlrobert/codegpt/predictions/TriggerCustomPredictionAction.kt
@@ -47,7 +47,7 @@ class TriggerCustomPredictionAction : EditorAction(Handler()), HintManagerImpl.A
             }
 
             ApplicationManager.getApplication().executeOnPooledThread {
-                service<PredictionService>().displayInlineDiff(editor, true)
+                service<PredictionService>().displayInlineDiff(editor)
             }
         }