+++ src/main/kotlin/ee/carlrobert/codegpt/settings/configuration/ConfigurationSettings.kt
@@ -23,7 +23,7 @@ class ConfigurationSettings :
 
 class ConfigurationSettingsState : BaseState() {
     var commitMessagePrompt by string(CoreActionsState.DEFAULT_GENERATE_COMMIT_MESSAGE_PROMPT)
-    var maxTokens by property(2048)
+    var maxTokens by property(8192)
     var temperature by property(0.1f) { max(0f, min(1f, it)) }
     var checkForPluginUpdates by property(true)
     var checkForNewScreenshots by property(true)
@@ -46,7 +46,6 @@ class ChatCompletionSettingsState : BaseState() {
 }
 
 class CodeCompletionSettingsState : BaseState() {
-    var multiLineEnabled by property(true)
     var treeSitterProcessingEnabled by property(true)
     var gitDiffEnabled by property(true)
     var collectDependencyStructure by property(true)