+++ src/main/kotlin/ee/carlrobert/codegpt/ui/textarea/PromptTextField.kt
@@ -6,6 +6,7 @@ import com.intellij.codeInsight.lookup.impl.PrefixChangeListener
 import com.intellij.ide.IdeEventQueue
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.application.runInEdt
+import com.intellij.openapi.application.runReadAction
 import com.intellij.openapi.application.runUndoTransparentWriteAction
 import com.intellij.openapi.components.service
 import com.intellij.openapi.editor.Editor
@@ -44,13 +45,14 @@ class PromptTextField(
     private val onTextChanged: (String) -> Unit,
     private val onBackSpace: () -> Unit,
     private val onLookupAdded: (LookupActionItem) -> Unit,
-    private val onSubmit: (String) -> Unit
+    private val onSubmit: (String) -> Unit,
 ) : EditorTextField(project, FileTypes.PLAIN_TEXT), Disposable {
 
-    private val coroutineScope = CoroutineScope(Dispatchers.Default + SupervisorJob())
+    private val coroutineScope = CoroutineScope(Dispatchers.Main + SupervisorJob())
     private var showSuggestionsJob: Job? = null
 
     val dispatcherId: UUID = UUID.randomUUID()
+    var lookup: LookupImpl? = null
 
     init {
         isOneLineMode = false
@@ -60,8 +62,14 @@ class PromptTextField(
 
     override fun onEditorAdded(editor: Editor) {
         IdeEventQueue.getInstance().addDispatcher(
-            PromptTextFieldEventDispatcher(dispatcherId, onBackSpace) {
+            PromptTextFieldEventDispatcher(dispatcherId, onBackSpace, lookup) {
+                val shown = lookup?.let { it.isShown && !it.isLookupDisposed } == true
+                if (shown) {
+                    return@PromptTextFieldEventDispatcher
+                }
+
                 onSubmit(text)
+                it.consume()
             },
             this
         )
@@ -95,14 +103,8 @@ class PromptTextField(
     }
 
     private fun showGroupLookup(editor: Editor, lookupElements: Array<LookupElement>) {
-        val lookup = project.service<LookupManager>().createLookup(
-            editor,
-            lookupElements,
-            "",
-            LookupArranger.DefaultArranger()
-        ) as LookupImpl
-
-        lookup.addLookupListener(object : LookupListener {
+        lookup = createLookup(editor, lookupElements, "")
+        lookup?.addLookupListener(object : LookupListener {
             override fun itemSelected(event: LookupEvent) {
                 val lookupString = event.item?.lookupString ?: return
                 val suggestion =
@@ -127,8 +129,8 @@ class PromptTextField(
                 }
             }
         })
-        lookup.refreshUi(false, true)
-        lookup.showLookup()
+        lookup?.refreshUi(false, true)
+        lookup?.showLookup()
     }
 
     private fun findAtSymbolPosition(editor: Editor): Int {
@@ -153,12 +155,14 @@ class PromptTextField(
         editor: Editor,
         lookupElements: Array<LookupElement>,
         searchText: String
-    ) = LookupManager.getInstance(project).createLookup(
-        editor,
-        lookupElements,
-        searchText,
-        LookupArranger.DefaultArranger()
-    ) as LookupImpl
+    ) = runReadAction {
+        LookupManager.getInstance(project).createLookup(
+            editor,
+            lookupElements,
+            searchText,
+            LookupArranger.DefaultArranger()
+        ) as LookupImpl
+    }
 
     private fun showSuggestionLookup(
         lookupElements: Array<LookupElement>,
@@ -166,8 +170,8 @@ class PromptTextField(
         filterText: String = "",
     ) {
         editor?.let {
-            val lookup = createLookup(it, lookupElements, filterText)
-            lookup.addLookupListener(object : LookupListener {
+            lookup = createLookup(it, lookupElements, filterText)
+            lookup?.addLookupListener(object : LookupListener {
                 override fun itemSelected(event: LookupEvent) {
                     val lookupItem = event.item?.getUserData(LookupItem.KEY) ?: return
                     if (lookupItem !is LookupActionItem) return
@@ -205,14 +209,14 @@ class PromptTextField(
                 }
             })
 
-            lookup.addPrefixChangeListener(object : PrefixChangeListener {
+            lookup?.addPrefixChangeListener(object : PrefixChangeListener {
                 override fun afterAppend(c: Char) {
                     showSuggestionsJob?.cancel()
                     showSuggestionsJob = coroutineScope.launch {
                         if (parentGroup is DynamicLookupGroupItem) {
                             val searchText = getSearchText()
                             if (searchText.length == 2) {
-                                parentGroup.updateLookupList(lookup, searchText)
+                                parentGroup.updateLookupList(lookup!!, searchText)
                             }
                         }
                     }
@@ -234,8 +238,8 @@ class PromptTextField(
 
             }, this)
 
-            lookup.refreshUi(false, true)
-            lookup.showLookup()
+            lookup?.refreshUi(false, true)
+            lookup?.showLookup()
         }
     }