+++ src/main/kotlin/ee/carlrobert/codegpt/ui/textarea/UserInputPanel.kt
@@ -277,14 +277,15 @@ class UserInputPanel(
         return when (service<GeneralSettings>().state.selectedService) {
             ServiceType.CUSTOM_OPENAI,
             ServiceType.ANTHROPIC,
+            ServiceType.GOOGLE,
             ServiceType.AZURE,
             ServiceType.OLLAMA -> true
 
             ServiceType.CODEGPT -> {
                 listOf(
-                    "gpt-4o",
-                    "gpt-4o-mini",
-                    "gemini-pro-1.5",
+                    "gpt-4.1",
+                    "gpt-4.1-mini",
+                    "gemini-pro-2.5",
                     "claude-3-opus",
                     "claude-3.5-sonnet",
                     "claude-3.7-sonnet"